deduplicate session manager benchmarks
