+++ runtime/mandala/src/benchmarking/mod.rs
@@ -63,8 +63,9 @@ pub mod prices {
 pub mod transaction_payment {
 	benchmarks::transaction_payment_benchmarks!();
 }
-
-pub mod session_manager;
+pub mod session_manager {
+	benchmarks::session_manager_benchmarks!();
+}
 
 // orml benchmarking
 pub mod auction {
